requirements.txt
transformers
langchain
openai
cohere
tiktoken
huggingface
python-dotenv
chardet
pandas